refactor: next migration, shared page구현, 링크 검색 기능 구현
