+++ src/app/contents.css
@@ -1,0 +1,18 @@
+.contents {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  gap: 2.5rem;
+  justify-content: flex-start;
+  padding: 2.5rem 0 6.25rem 0;
+  width: auto;
+}
+
+.card_list_container {
+  position: releative;
+  display: flex;
+  flex-direction: column;
+  gap: 2.5rem;
+  justify-content: space-between;
+}
