+++ src/app/folder/(folderContents)/folderContents.css
@@ -39,6 +39,10 @@
   background: #fff;
 }
 
+.folder_group_item > * {
+  text-decoration: none;
+}
+
 .folder_toggle {
   display: flex;
   padding: 0.5rem 0.75rem;
@@ -49,6 +53,10 @@
   border-radius: 0.3125rem;
   border: 1px solid var(--Linkbrary-primary-color, #6d6afe);
   background: var(--Linkbrary-primary-color, #6d6afe);
+}
+
+.folder_toggle > * {
+  color: #ffffff;
 }
 
 .folder_group_title {
