+++ src/app/folder/[id]/layout.tsx
