+++ src/app/shared/page.tsx
@@ -1,17 +1,51 @@
-import SharedContents from "../../components/contents/sharedContents/SharedContents";
-import SharedHeader from "../../components/header/SharedHeader";
+import SharedContents from "./(sharedContents)/SharedContents";
+import SharedHeader from "./(sharedContents)/SharedHeader";
 // import NavigatorBar from "../../components/navigatorBar/NavigatorBar";
 
-export default function Shared({
-  params: { userParam, folderParam },
+import { getFolderInfo, getFolderGroup, getSelectionFolderLinks, setFolderLinksFromItems } from "../../api/api";
+
+const getFolderLink = async (userParam: number, folderParam: string) => {
+  if (userParam == undefined || folderParam == undefined) return [];
+
+  const response = await getSelectionFolderLinks(folderParam, userParam);
+
+  if (response.data == undefined) return [];
+  return setFolderLinksFromItems(response.data);
+};
+
+const getFolderName = async (userParam: number, folderParam: string) => {
+  if (userParam == undefined || folderParam == undefined) return "";
+
+  const result = await getFolderGroup(userParam);
+  if (result.data == undefined) return "";
+
+  const items = result.data;
+  const findItem = items.find((item: any) => String(item.id) == String(folderParam)); // FixME: type any
+
+  return findItem == undefined ? "" : findItem.name;
+};
+
+const getFolder = async () => {
+  const result = await getFolderInfo();
+
+  return result.owner;
+};
+
+export default async function Shared({
+  searchParams: { user, folder },
 }: {
-  params: { userParam: number; folderParam: string };
+  searchParams: { user: number; folder: string };
 }) {
+  const folderName: string = await getFolderName(user, folder);
+  const folderInfo: any = await getFolder();
+
+  const folderLinks = await getFolderLink(user, folder);
+
   return (
     <>
       {/* <NavigatorBar /> */}
-      <SharedHeader userParam={userParam} folderParam={folderParam} />
-      <SharedContents userParam={userParam} folderParam={folderParam} />
+      <SharedHeader folderName={folderName} folderInfo={folderInfo} />
+      <SharedContents folderLinks={folderLinks} />
     </>
   );
 }
