+++ src/components/contents/searchBar/searchBar.css
@@ -1,13 +1,4 @@
-.contents {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  gap: 2.5rem;
-  justify-content: flex-start;
-  padding: 2.5rem 0 6.25rem 0;
-  width: auto;
-}
+
 
 .search_bar {
   align-items: flex-start;
@@ -35,12 +26,4 @@
 
 .search_delete:hover {
   filter: brightness(0.8);
-}
-
-.card_list_container {
-  position: releative;
-  display: flex;
-  flex-direction: column;
-  gap: 2.5rem;
-  justify-content: space-between;
 }
